Reorganize implementation slightly---organizing as one file now.
